refactor(v1): clean files name
