+++ src/components/furniture/Hardware.vue
