+++ src/components/furniture/Separators.vue
