+++ src/router.js
@@ -1,9 +1,9 @@
 import { createWebHistory, createRouter } from 'vue-router';
 
 import ColorFurniture from './pages/ColorFurniture.vue';
-import Separators from './pages/Separators.vue';
+import AddSeparators from './pages/AddSeparators.vue';
 import CreateFurniture from './pages/CreateFurniture.vue';
-import Hardware from './pages/Hardware.vue';
+import AddHardware from './pages/AddHardware.vue';
 import Reference from './pages/Reference.vue';
 import Home from './pages/Home.vue';
 import NotFound from './pages/NotFound.vue';
@@ -14,8 +14,8 @@ const router = createRouter({
     { path: '/', component: Home },
     { path: '/reference', component: Reference },
     { path: '/create-furniture', component: CreateFurniture },
-    { path: '/separators', component: Separators },
-    { path: '/hardware', component: Hardware },
+    { path: '/add-separators', component: AddSeparators },
+    { path: '/add-hardware', component: AddHardware },
     { path: '/color', component: ColorFurniture },
     { path: '/:notFound(.*)', component: NotFound },
   ],
